This changeset includes an initial enhancement per ticket:667. The reduction logic for the NormalEquation reduction is now dynamic and should do a reasonable optimal two stage reduction.
Also includes some fixes to the test scripts.


git-svn-id: https://svn.atnf.csiro.au/askapsdp/trunk@2046 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/CP/imager/trunk/functests/continuum-cycling/run.sh
@@ -20,6 +20,11 @@ mv 10uJy_stdtest.ms 10uJy_stdtest_1.ms
 echo Done
 
 mpirun -np 2 ${ASKAP_ROOT}/Code/Components/CP/imager/trunk/apps/imager.sh -inputs clean.in | tee $OUTPUT
+if [ $? -ne 0 ]; then
+    echo Error: mpirun returned an error
+    exit 1
+fi
+
 
 echo -n Removing measurement set...
 rm -rf 10uJy_stdtest_0.ms
