+++ Code/Components/CP/imager/trunk/functests/continuum/run.sh
@@ -20,6 +20,10 @@ mv 10uJy_stdtest.ms 10uJy_stdtest_1.ms
 echo Done
 
 mpirun -np 2 ${ASKAP_ROOT}/Code/Components/CP/imager/trunk/apps/imager.sh -inputs dirty.in | tee $OUTPUT
+if [ $? -ne 0 ]; then
+    echo Error: mpirun returned an error
+    exit 1
+fi
 
 echo -n Removing measurement set...
 rm -rf 10uJy_stdtest_0.ms
