+++ Code/Components/CP/imager/trunk/functests/spectralline/run.sh
@@ -2,7 +2,7 @@
 
 export AIPSPATH=${ASKAP_ROOT}/Code/Components/Synthesis/testdata/trunk
 
-if [ ! -x ${ASKAP_ROOT}/Code/Components/CP/imager/branches/specline-dev/apps/imager.sh ]; then
+if [ ! -x ${ASKAP_ROOT}/Code/Components/CP/imager/trunk/apps/imager.sh ]; then
     echo imager.sh does not exit
 fi
 
@@ -17,7 +17,11 @@ tar zxf ../10uJy_stdtest.ms.tgz
 mv 10uJy_stdtest.ms 10uJy_stdtest_1.ms
 echo Done
 
-mpirun -np 2 ${ASKAP_ROOT}/Code/Components/CP/imager/branches/specline-dev/apps/imager.sh -inputs imager.in
+mpirun -np 2 ${ASKAP_ROOT}/Code/Components/CP/imager/trunk/apps/imager.sh -inputs imager.in
+if [ $? -ne 0 ]; then
+    echo Error: mpirun returned an error
+    exit 1
+fi
 
 echo -n Removing measurement set...
 rm -rf 10uJy_stdtest_0.ms
